+layout.svelteに移動できそうな変数を全部移動する
